+++ Editor/SettingsInspector.cs
@@ -144,6 +144,9 @@ namespace UnityBHL
       var forceOnPlayProp = serializedObject.FindProperty(nameof(Settings.forceRecompileOnPlay));
       EditorGUILayout.PropertyField(forceOnPlayProp, new GUIContent("Recompile On Play", forceOnPlayProp.tooltip));
 
+      var hotReloadProp = serializedObject.FindProperty(nameof(Settings.hotReloadOnRecompile));
+      EditorGUILayout.PropertyField(hotReloadProp, new GUIContent("Hot Reload On Recompile", hotReloadProp.tooltip));
+
       serializedObject.ApplyModifiedProperties();
     }